FormControl tests
This sets up a place to test form text fields that will automatically
validate against the server. It makes tests around the existing text
field functionality without the auto-server-validation.

I pulled out the DeepMind fake model from an existing test file in order
to share it with this new test file.
